+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 2.9.12 (2023-06-06)
 
 * (theimo1221) Move direction setting from Window to the Shutter, to prevent group-settings need
 * (theimo1221) Make heat reduction position controllable